warehouse/pyproject.toml
[project]
name = "warehouse"
authors = [{name = "Nikita Khirny", email = "khirnick@yandex.ru"}]
description = "Warehouse"
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
    "tortoise-orm==0.22.1"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "ruff"
]
dev = ["ipython"]

[tool.pytest.ini_options]
addopts = "--cov-report term-missing --cov-report xml --cov src/"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
